+++ uspace/lib/posix/include/posix/sys/time.h
@@ -33,6 +33,22 @@
 #ifndef POSIX_SYS_TIME_H_
 #define POSIX_SYS_TIME_H_
 
+#define DST_NONE  0
+
+typedef long suseconds_t;
+
+struct timeval {
+	time_t tv_sec;        /* seconds */
+	suseconds_t tv_usec;  /* microseconds */
+};
+
+struct timezone {
+	int tz_minuteswest;  /* minutes W of Greenwich */
+	int tz_dsttime;      /* type of dst correction */
+};
+
+extern int gettimeofday(struct timeval *, struct timezone *);
+
 #endif
 
 /** @}
